data binder

+++ Assets/Scripts/Player/PlayerSharedConfigData.cs
@@ -3,5 +3,5 @@
 [CreateAssetMenu(fileName = nameof(PlayerSharedConfigData), menuName = nameof(PlayerSharedConfigData))]
 public class PlayerSharedConfigData : ScriptableObject
 {
-    public float maxSpeed;
+    public int maxSpeed;
 }
